requirements.txt
edge-tts>=6.1.7
moviepy>=1.0.3
python-dotenv>=1.0.0
numpy>=1.24.0
pydub>=0.25.1